requirements.txt
Django==3.2.13
djangorestframework==3.13.1
celery==5.2.6
redis==4.2.2
